+++ src/components/Snipcart.jsx
