+++ src/components/SnipcartProvider.js
